Adicionando interfaces para os use cases, adicionando interface de usecases nas configurações, e refatorando código
