+++ src/main/java/br/com/fiap/productmanagement/domain/usecase/PrepareProductToProcessUseCase.java
@@ -1,0 +1,6 @@
+package br.com.fiap.productmanagement.domain.usecase;
+
+import br.com.fiap.productmanagement.domain.entities.ProductEntity;
+import org.springframework.batch.item.ItemProcessor;
+
+public interface PrepareProductToProcessUseCase extends ItemProcessor<ProductEntity, ProductEntity> {}
